requirements.txt
matplotlib==3.5.2
networkx==2.5.1
numpy==1.23.0
pandas==1.4.3
scikit_network==0.26.0
snap_stanford==6.0.0
